pyproject.toml
[tool.poetry]
name = "salduba_corvino"
version = "0.1.0"
description = "Automated Trade Management for Corvino Fund"
authors = ["Miguel Pinilla <miguel.pinilla@saldubatech.com>"]
readme = ["README.md"]
repository = "https://github.com/Saldubatech/poetry-sample"
license = "Proprietary"
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Typing :: Typed"
]
packages = [
  {include = "salduba", from = "src"}
]
exclude = ["ib_src"]

[tool.poetry.dependencies]
python = "^3.11"

ibapi = "^9.81.1.post1"
pandas = "^2.2.2"
pyyaml = "^6.0.1"
[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
sphinx = "^7.3.7"
mypy = "^1.10.0"
pytest-cov = "^5.0.0"
pre-commit = "^3.7.1"
flake8 = "^7.0.0"


# https://github.com/jvllmr/poetry-types
isort = "^5.13.2"
black = "^24.4.2"

[tool.poetry.group.types.dependencies]
types-flake8 = "^7.0.0.20240511"

types-pyyaml = "^6.0.12.20240311"
pandas-stubs = "^2.2.2.240514"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
# This is what defines what "top-level" scripts will be published.
verify_contracts = "salduba.corvino.commands.verify_contracts:verify_contracts"

[tool.pytest.ini_options]
addopts = "--cov=salduba"

[tool.mypy]
strict=true
python_version = "3.11"
exclude = "ib_src"

[[tool.mypy.overrides]]
module = 'ibapi.*'
ignore_missing_imports = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 130
skip_gitignore = true
# skip_glob = ["tests/data", "profiling"]
# known_first_party = ["black", "blib2to3", "blackd", "_black_version"]

[tool.black]
line-length = 130
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
